# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DuckOJ API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/submissions": {
      "post": {
        "summary": "Submit a solution for grading",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "problemCode": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "languageKey": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 32
                  },
                  "source": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 65536
                  }
                },
                "required": [
                  "problemCode",
                  "languageKey",
                  "source"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The submission was accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "default": "about:blank"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "code"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No such problem, or no such language",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "default": "about:blank"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "code"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "The request body failed validation",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "default": "about:blank"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "code"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{id}": {
      "get": {
        "summary": "A submission visible to the caller",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "The submission",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "problemCode": {
                      "type": "string"
                    },
                    "languageKey": {
                      "type": "string"
                    },
                    "state": {
                      "type": "string",
                      "enum": [
                        "queued",
                        "compiling",
                        "grading",
                        "done",
                        "errored"
                      ]
                    },
                    "verdict": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "enum": [
                        "AC",
                        "WA",
                        "TLE",
                        "MLE",
                        "OLE",
                        "RTE",
                        "IR",
                        "IE",
                        null
                      ]
                    },
                    "points": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "maxPoints": {
                      "type": [
                        "number",
                        "null"
                      ]
                    },
                    "timeMs": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "memoryKb": {
                      "type": [
                        "integer",
                        "null"
                      ]
                    },
                    "compileOutput": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "cases": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "groupIndex": {
                            "type": "integer"
                          },
                          "caseIndex": {
                            "type": "integer"
                          },
                          "verdict": {
                            "type": [
                              "string",
                              "null"
                            ],
                            "enum": [
                              "AC",
                              "WA",
                              "TLE",
                              "MLE",
                              "OLE",
                              "RTE",
                              "IR",
                              "IE",
                              null
                            ]
                          },
                          "skipped": {
                            "type": "boolean"
                          },
                          "timeMs": {
                            "type": "integer"
                          },
                          "memoryKb": {
                            "type": "integer"
                          },
                          "points": {
                            "type": "number"
                          },
                          "maxPoints": {
                            "type": "number"
                          },
                          "feedback": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        },
                        "required": [
                          "groupIndex",
                          "caseIndex",
                          "verdict",
                          "skipped",
                          "timeMs",
                          "memoryKb",
                          "points",
                          "maxPoints",
                          "feedback"
                        ]
                      }
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "judgedAt": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "id",
                    "problemCode",
                    "languageKey",
                    "state",
                    "verdict",
                    "points",
                    "maxPoints",
                    "timeMs",
                    "memoryKb",
                    "compileOutput",
                    "cases",
                    "createdAt",
                    "judgedAt"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "default": "about:blank"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "code"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "No such submission, or one the caller may not see — the two are indistinguishable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "default": "about:blank"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "code"
                  ]
                }
              }
            }
          },
          "422": {
            "description": "The `id` path parameter is not a valid submission id",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "default": "about:blank"
                    },
                    "title": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "code"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}